Changed folder structure
Updated utils.py to helper.py
